Remove DateRangePicker and fix bugs.
