--- app/src/androidTest/java/me/adawoud/bottomsheetpickers/ExampleInstrumentedTest.kt
@@ -1,22 +1,0 @@
-package me.adawoud.bottomsheetpickers
-
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-import org.junit.Assert.assertEquals
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("me.adawoud.bottomsheetpickers", appContext.packageName)
-    }
-}
