--- app/src/main/res/values/dimens.xml
@@ -1,4 +1,0 @@
-<resources>
-    <dimen name="fab_margin">16dp</dimen>
-    <dimen name="spacing_default">16dp</dimen>
-</resources>
