--- bottomsheetrangepickers/publish.gradle
@@ -1,54 +1,0 @@
-apply plugin: 'maven-publish'
-apply plugin: 'com.jfrog.bintray'
-
-version '0.94'
-group 'com.adawoud'
-
-publishing {
-    publications {
-        Production(MavenPublication) {
-            artifact("$buildDir/outputs/aar/bottomsheetrangepickers-release.aar")
-            groupId this.group
-            artifactId 'bottomsheetrangepickers'
-            version this.version
-
-            pom.withXml {
-                def dependenciesNode = asNode().appendNode('dependencies')
-
-                // Iterate over the implementation dependencies (we don't want the test ones),
-                // adding a <dependency> node for each
-                configurations.implementation.allDependencies.each {
-                    // Ensure dependencies such as fileTree are not included in the pom.
-                    if (it.name != 'unspecified') {
-                        def dependencyNode = dependenciesNode.appendNode('dependency')
-                        dependencyNode.appendNode('groupId', it.group)
-                        dependencyNode.appendNode('artifactId', it.name)
-                        dependencyNode.appendNode('version', it.version)
-                    }
-                }
-            }
-        }
-    }
-}
-
-bintray {
-    user = project.hasProperty('user') ?: System.getenv('BINTRAY_USER')
-    key =  project.hasProperty('apiKey') ?: System.getenv('BINTRAY_API_KEY')
-    publications = ['Production']
-    configurations = ['archives']
-    // Default: false. Whether to run this as dry-run, without deploying
-    dryRun = false
-    // Default: false. Whether to override version artifacts already published
-    override = false
-    // Default: false. Whether version should be auto published after an upload
-    publish = true
-    pkg {
-        repo = 'BottomSheetRangePickers'
-        name = 'com.adawoud.bottomsheetrangepickers'
-        version {
-            name = this.version
-            released = new Date()
-            vcsTag = this.version
-        }
-    }
-}
