+++ bottomsheettimerangepicker/publish.gradle
@@ -1,0 +1,54 @@
+apply plugin: 'maven-publish'
+apply plugin: 'com.jfrog.bintray'
+
+version '1.0'
+group 'com.adawoud'
+
+publishing {
+    publications {
+        Production(MavenPublication) {
+            artifact("$buildDir/outputs/aar/bottomsheettimerangepicker-release.aar")
+            groupId this.group
+            artifactId 'bottomsheettimerangepicker'
+            version this.version
+
+            pom.withXml {
+                def dependenciesNode = asNode().appendNode('dependencies')
+
+                // Iterate over the implementation dependencies (we don't want the test ones),
+                // adding a <dependency> node for each
+                configurations.implementation.allDependencies.each {
+                    // Ensure dependencies such as fileTree are not included in the pom.
+                    if (it.name != 'unspecified') {
+                        def dependencyNode = dependenciesNode.appendNode('dependency')
+                        dependencyNode.appendNode('groupId', it.group)
+                        dependencyNode.appendNode('artifactId', it.name)
+                        dependencyNode.appendNode('version', it.version)
+                    }
+                }
+            }
+        }
+    }
+}
+
+bintray {
+    user = project.hasProperty('user') ?: System.getenv('BINTRAY_USER')
+    key = project.hasProperty('apiKey') ?: System.getenv('BINTRAY_API_KEY')
+    publications = ['Production']
+    configurations = ['archives']
+    // Default: false. Whether to run this as dry-run, without deploying
+    dryRun = false
+    // Default: false. Whether to override version artifacts already published
+    override = true
+    // Default: false. Whether version should be auto published after an upload
+    publish = true
+    pkg {
+        repo = 'BottomSheetTimeRangePicker'
+        name = 'com.adawoud.bottomsheettimerangepicker'
+        version {
+            name = this.version
+            released = new Date()
+            vcsTag = this.version
+        }
+    }
+}
