+++ sample/src/main/res/values/dimens.xml
@@ -1,0 +1,3 @@
+<resources>
+    <dimen name="spacing_default">16dp</dimen>
+</resources>
