+++ sample/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+    <string name="app_name">BottomSheetPickers</string>
+    <string name="time_range_picker">Time range picker</string>
+    <string name="time_range">Time range</string>
+    <string name="chosen_time_range">%1$s:%2$s - %3$s:%4$s</string>
+</resources>
