+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':bottomsheetrangepickers'
+include ':sample', ':bottomsheettimerangepicker'
